% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassIVE_functions.R
\name{MassIVE-utils}
\alias{MassIVE-utils}
\alias{massive_ftp_path}
\alias{massive_list_files}
\alias{massive_download_file}
\alias{massive_param_file}
\alias{massive_number_files}
\alias{massive_sync_data_files}
\alias{massive_cached_data_files}
\alias{massive_delete_cache}
\title{Utility functions for the MassIVE repository}
\usage{
massive_ftp_path(x = character(), mustWork = TRUE)

massive_list_files(x = character(), pattern = NULL)

massive_download_file(
  massiveId = character(),
  pattern = "*",
  fileName = character(),
  path = "./",
  overwrite = FALSE
)

massive_param_file(massiveId = character(), fileName = "params.xml")

massive_number_files(
  massiveId = character(),
  pattern = "mzML$|CDF$|cdf$|mzXML$"
)

massive_sync_data_files(
  massiveId = character(),
  pattern = "mzML$|CDF$|cdf$|mzXML$",
  fileName = character()
)

massive_cached_data_files(
  massiveId = character(),
  pattern = "*",
  fileName = character()
)

massive_delete_cache(massiveId = character())
}
\arguments{
\item{x}{\code{character(1)} with the ID of the MassIVE data set (usually
starting with a \emph{MSV} followed by a number).}

\item{mustWork}{for \code{massive_ftp_path()}: \code{logical(1)} whether the validity
of the path should be verified or not. By default (with
\code{mustWork = TRUE}) the function throws an error if either the data set
does not exist or if the folder can not be accessed (e.g. if no internet
connection is available).}

\item{pattern}{for \code{massive_list_files()}, \code{massive_sync_data_files()},
\code{massive_cached_data_files()}, \code{massive_donwload_file()}, and
\code{massive_number_files()}: \code{character(1)} defining a pattern to filter
the file names, such as \code{pattern = "mzML$"} to retrieve the file names
of all files of the data set (i.e., files with extension \code{"mzML"}). This
parameter is passed to the \code{\link[=grepl]{grepl()}} function.}

\item{massiveId}{\code{character(1)} with the ID of a single MassIVE data
set/experiment.}

\item{fileName}{for \code{massive_sync_data_files()},
\code{massive_cached_data_files()} and \code{massive_download_file()}: optional
\code{character} defining the names of specific data files of a data set that
should be downloaded and cached.}

\item{path}{for \code{massive_download_file()}: optional \code{character} defining the
directory where download the files.}

\item{overwrite}{for \code{massive_download_file()}: \code{logical(1)} whether
existing files should be overwritten. Defaults to \code{FALSE}, in which
case files that already exist in \code{path} are skipped.}
}
\value{
\itemize{
\item For \code{massive_ftp_path()}: \code{character(1)} with the ftp path to the specified
data set on the MassIVE ftp server.
\item For \code{massive_list_files()}: \code{character} with the names of the files in the
data set's base ftp directory.
\item For \code{massive_sync_data_files()} and \code{massive_cached_data_files()}: a
\code{data.frame} with the MassIVE ID, the name(s) and remote and
local file names of the synchronized data files sorted by filename
(\code{"rpath"}).
\item For \code{massive_number_files()}: \code{integer(1)} with the number of data files
in the data set.
}
}
\description{
\href{https://massive.ucsd.edu/ProteoSAFe/static/massive.jsp}{MassIVE}
(Mass Spectrometry Interactive Virtual Environment) is a community
resource developed by the NIH-funded Center for Computational Mass
Spectrometry to promote the global, free exchange of mass spectrometry
data. MassIVE supports deposition of both proteomics and metabolomics
experiments, and is a full member of the
\href{http://www.proteomexchange.org/}{ProteomeXchange} consortium, allowing
datasets to be assigned ProteomeXchange accessions to satisfy publication
requirements. Submitted data can include raw mass spectrometry files,
identification results, and quantification data. The repository also
provides online workflows for reanalysis of public datasets and tools
for comparison of identification results across datasets.

Each experiment in MassIVE is identified with its unique identifier,
starting with \emph{MSV} followed by a number. The data (raw MS files,
metadata, and result files) of a dataset are available for public
download and online browsing once the dataset has been made public by
its submitter.

The functions listed here allow to query and retrieve information of a
data set/experiment from MassIVE.
\itemize{
\item \code{massive_ftp_path()}: returns the FTP path for a provided MassIVE ID.
If the MassIVE ID does not exist the function throws an error.
With \code{mustWork = TRUE} (the default) the function throws an error
either because the data set does not exist in
\href{https://datasetcache.gnps2.org/datasette/database/filename}{GNPS2 DB} (No
mzML/CDF/mzXML files available) or no internet connection is available.
The function returns a \code{character(1)} with the FTP path to the data set
folder.
\item \code{massive_cached_data_files()}: lists locally cached data files from
MassIVE. Since this function evaluates only local content it does not
require an internet connection. With the default parameters all available
data files are listed. The parameters can be used to restrict the lookup.
\item \code{massive_list_files()}: returns the available files (and directories) for
the specified MassIVE data set (i.e., the FTP directory content of the
data set). The function returns a \code{character} vector with the relative
file names to the absolute FTP path (\code{massive_ftp_path()}) of the data set.
Parameter \code{pattern} allows to filter the file names and define which
file names should be returned.
\item \code{massive_sync_data_files()}: synchronize data files of a specified
MassIVE data set eventually downloading and locally caching them.
Parameter \code{fileName} allows to specify names of selected data files to
sync.
\item \code{massive_download_file()}: download files from the MassIVE repository for a
specified MassIVE dataset. Use \code{pattern} to filter files by name using a
regular expression (downloads all files by default). Use \code{fileName} to
specify one or more exact file names to download. Use \code{path} to set the
destination directory for downloaded files.
\item \code{massive_param_file()}: download and parse the \code{params.xml} files of
the data set. The function return a \code{data.frame} or a \code{list} of
\code{data.frame} with 2 columns (Parameter Name, Value). Use \code{fileName} to
parse additional \code{xml} files in the data.set.
\item \code{massive_number_files()}: return the number of data files in a specified
MassIVE data set. Use \code{pattern} to filter files by name using a regular
expression, default: \code{pattern = "mzML$|CDF$|cdf$|mzXML$"}.
\item \code{massive_delete_cache()}: removes all local content for the MassIVE
data set with ID \code{massiveId}. This will delete eventually present
locally cached data files for the specified data set. This does not
change any other data eventually present in the local \code{BiocFileCache}.
}
}
\details{
Data retrieval follows three main steps. First, the package queries the
\href{https://datasetcache.gnps2.org/datasette/database/filename}{GNPS2 DB}
to list all files for the provided \code{massiveId}, filtering them by
\code{filePattern} to retain only formats supported by \code{MsBackendMzR} (mzML,
CDF, mzXML). Second, the FTP link is retrieved from
\href{https://massive.ucsd.edu/ProteoSAFe/}{MassIVE}. If the requested files are
in the \code{ccms_peak} folder, the FTP link is updated by changing the volume
from the project-specific one to volume \code{z01}, which contains the
\code{ccms_peak} folder for all projects. Each file is then downloaded from the
MassIVE FTP server and cached locally. Files already present in the cache
are not re-downloaded. Third, the cached local paths are passed to
\code{\link[Spectra:MsBackendMzR]{Spectra::MsBackendMzR()}} to read and index the spectral data. Two
additional per-spectrum variables are populated: \code{"massive_id"} and
\code{"data_file"}. When \code{offline = TRUE}, the remote query is skipped and
only previously cached content is used.
}
\examples{

## Get the FTP path to the data set MSV000080547
massive_ftp_path("MSV000080547")

## Retrieve available files (and directories) for the data set MSV000080547
massive_list_files("MSV000080547")

## Retrieve the available .mzML files.
mzMLfiles <- massive_list_files("MSV000080547", pattern = "mzML$")
mzMLfiles

## Download parameter file for the data set MSV000080547
massive_download_file("MSV000080547", pattern = "params.xml",
                      path = tempdir())

}
\author{
Johannes Rainer, Philippine Louail, Gabriele Tomè
}
