% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendMassIVE.R
\name{MsBackendMassIVE}
\alias{MsBackendMassIVE}
\alias{MsBackendMassIVE-class}
\alias{backendInitialize,MsBackendMassIVE-method}
\alias{backendRequiredSpectraVariables,MsBackendMassIVE-method}
\alias{massive_sync}
\title{MsBackend representing MS data from MassIVE}
\usage{
MsBackendMassIVE()

\S4method{backendInitialize}{MsBackendMassIVE}(
  object,
  massiveId = character(),
  filePattern = "mzML$|CDF$|cdf$|mzXML$",
  offline = FALSE,
  ...
)

\S4method{backendRequiredSpectraVariables}{MsBackendMassIVE}(object, ...)

massive_sync(x, offline = FALSE)
}
\arguments{
\item{object}{an instance of \code{MsBackendMassIVE}.}

\item{massiveId}{\code{character(1)} with the ID of a single MassIVE data
set/experiment.}

\item{filePattern}{\code{character} with the pattern defining the supported (or
requested) file types. Defaults to
\code{filePattern = "mzML$|CDF$|cdf$|mzXML$"} hence restricting to mzML,
CDF and mzXML files which are supported by \emph{Spectra}'s
\code{MsBackendMzR} backend.}

\item{offline}{\code{logical(1)} whether only locally cached content should be
evaluated/loaded.}

\item{...}{additional parameters; currently ignored.}

\item{x}{an instance of \code{MsBackendMassIVE}.}
}
\value{
\itemize{
\item For \code{MsBackendMassIVE()}: an instance of \code{MsBackendMassIVE}.
\item For \code{backendInitialize()}: an instance of \code{MsBackendMassIVE} with
the MS data of the specified MassIVE data set.
\item For \code{backendRequiredSpectraVariables()}: \code{character} with spectra
variables that are needed for the backend to provide the MS data.
\item For \code{massive_sync()}: the input \code{MsBackendMassIVE} with the paths to
the locally cached data files being eventually updated.
}
}
\description{
\code{MsBackendMassIVE} retrieves and represents mass spectrometry (MS)
data from proteomics and metabolomics experiments stored in the
\href{https://massive.ucsd.edu/ProteoSAFe/static/massive.jsp}{MassIVE}
(Mass Spectrometry Interactive Virtual Environment) repository, a
community resource developed by the NIH-funded Center for Computational
Mass Spectrometry at UC San Diego. The backend directly extends the
\link[Spectra:MsBackend]{Spectra::MsBackendMzR} backend from the \emph{Spectra} package and hence
supports MS data in mzML, netCDF and mzXML format. Data in other formats
can not be loaded with \code{MsBackendMassIVE}.
Upon initialization with the \code{backendInitialize()} method, the
\code{MsBackendMassIVE} backend downloads and caches the MS data files of
a dataset locally, avoiding repeated download of the data.
The local data cache is managed by Bioconductor's \emph{BiocFileCache} package.
See the help and vignettes from that package for details on cached data
resources. Additional utility functions for management of cached files are
also provided by \emph{MsBackendMassIVE}. See help for
\code{\link[=massive_cached_data_files]{massive_cached_data_files()}} for more information.
}
\details{
File names for data files are by default extracted from the column
\code{"filepath"} of the
\href{https://datasetcache.gnps2.org/datasette/database/filename}{GNPS2 database}.

The backend uses the
\href{https://bioconductor.org/packages/BiocFileCache}{BiocFileCache} package for
caching of the data files. These are stored in the default local
\emph{BiocFileCache} cache along with additional metadata that includes the
MassIVE ID. Note that at present only MS data files in \emph{mzML}, \emph{CDF}
and \emph{mzXML} format are supported.

The \code{MsBackendMassIVE} backend defines and provides additional spectra
variables \code{"massive_id"} and \code{"data_file"} that list the MassIVE ID,
and the original data file name on the MassIVE ftp
server for each individual spectrum. The \code{"data_file"} can
be used for the mapping between the experiment's samples and the
individual data files, respective their spectra.

The \code{MsBackendMassIVE} backend is considered \emph{read-only} and does
thus not support changing \emph{m/z} and intensity values directly.
}
\note{
To account for high server load and eventually failing or rejected
downloads from the MassIVE FTP server (\verb{ftp://massive-ftp.ucsd.edu/}), the
download functions repeatedly retry to download a file. An error is thrown
if the download fails for 5 consecutive attempts. Between each attempt,
the function waits for an increasing time period (5 seconds between the
first and second and 10 seconds between the 2nd and 3rd attempt). This
time period can also be configured with the \code{"massive.sleep_mult"} option,
which defines the \emph{sleep time multiplicator} (defaults to 5).
}
\section{Initialization and loading of data}{


New instances of the class can be created with the \code{MsBackendMassIVE()}
function. Data is loaded and initialized using the \code{backendInitialize()}
function which can be configured with parameters \code{massiveId} and
\code{filePattern}. \code{massiveId} must be the ID of a \strong{single} (existing)
MassIVE dataset (e.g. \code{"MSV000079514"}). Optional parameter \code{filePattern}
defines the pattern used to filter the file names of the MS data files.
It defaults to data files with file endings of supported MS data formats.
\code{backendInitialize()} requires an active internet connection as the
function first compares the remote file content to the locally cached
files and eventually synchronizes changes/updates. This can be skipped
with \code{offline = TRUE} in which case only locally cached content is queried.

The \code{backendRequiredSpectraVariables()} function returns the names of the
spectra variables required for the backend to provide the MS data.

The \code{massive_sync()} function can be used to \emph{synchronize} the local data
cache and ensure that all data files are locally available. The function
will check the local cache and eventually download missing data files from
the MassIVE repository.
}

\examples{

library(MsBackendMassIVE)

## List files of a MassIVE data set
massive_list_files("MSV000080547")

## Initialize a MsBackendMassIVE representing all MS data files of
## the data set with the ID "MSV000080547". This will download and cache all
## files and subsequently load and represent them in R.

be <- backendInitialize(MsBackendMassIVE(), "MSV000080547",
                        filePattern = "1.mzML$")
be

## The `massive_sync()` function can be used to ensure that all data files
## are available locally. This function will eventually download missing data
## files or update their paths.
be <- massive_sync(be)
}
\author{
Gabriele Tomè, Philippine Louail, Johannes Rainer
}
